+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADAppManagementPolicy/MSFT_AADAppManagementPolicy.psm1
@@ -6,35 +6,23 @@ function Get-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $DisplayName,
 
         [Parameter()]
+        [System.String]
+        $Id,
+
+        [Parameter(Mandatory = $true)]
         [System.String]
         $Description,
 
         [Parameter()]
-        [System.String[]]
-        $AppPresetList,
-
-        [Parameter()]
-        [System.String[]]
-        $AppPresetMeetingList,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedAppBarApps,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedMessageBarApps,
-
-        [Parameter()]
         [System.Boolean]
-        $AllowUserPinning,
+        $IsEnabled,
 
         [Parameter()]
-        [System.Boolean]
-        $AllowSideLoading,
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $Restrictions,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -66,7 +54,7 @@ function Get-TargetResource
         $AccessTokens
     )
 
-    New-M365DSCConnection -Workload 'MicrosoftTeams' `
+    New-M365DSCConnection -Workload 'MicrosoftGraph' `
         -InboundParameters $PSBoundParameters | Out-Null
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -85,46 +73,69 @@ function Get-TargetResource
     $nullResult.Ensure = 'Absent'
     try
     {
-        $instance = Get-CsTeamsAppSetupPolicy -Identity $Identity -ErrorAction SilentlyContinue
+        if ($null -ne $Script:exportedInstances -and $Script:ExportMode)
+        {
+            $instance = $Script:exportedInstances | Where-Object -FilterScript {$_.Id -eq $Id}
+        }
+        else
+        {
+            if (-not [System.String]::IsNullOrEmpty($Id))
+            {
+                $instance = Get-MgBetaPolicyAppManagementPolicy -AppManagementPolicyId $Id `
+                                                                -ErrorAction SilentlyContinue
+            }
+            else
+            {
+                $instance = Get-MgBetaPolicyAppManagementPolicy | Where-Object -FilterScript {$_.DisplayName -eq $DisplayName}
+            }
+
+        }
         if ($null -eq $instance)
         {
             return $nullResult
         }
 
-        $AppPresetListValue = $instance.AppPresetList.Id
-        if ($instance.AppPresetList.Count -eq 0)
-        {
-            $AppPresetListValue = @()
+        $restrictionsValue = @{
+            passwordCredentials     = @()
+            keyCredentials          = @()
         }
 
-        $AppPresetMeetingListValue = $instance.AppPresetMeetingList.Id
-        if ($instance.AppPresetMeetingList.Count -eq 0)
+        foreach ($passwordCred in $instance.Restrictions.PasswordCredentials)
         {
-            $AppPresetMeetingListValue = @()
+            $newItem = @{
+                restrictForAppsCreatedAfterDateTime = $passwordCred.RestrictForAppsCreatedAfterDateTime.ToString()
+                restrictionType                     = $passwordCred.RestrictionType
+                state                               = $passwordCred.State
+            }
+            if ($null -ne $passwordCred.MaxLifetime)
+            {
+                $iso8601Duration = "P{0}DT{1}H{2}M{3}S" -f $passwordCred.MaxLifetime.Days, $passwordCred.MaxLifetime.Hours, $passwordCred.MaxLifetime.Minutes, $passwordCred.MaxLifetime.Seconds
+                $newItem.Add('maxLifetime', $iso8601Duration)
+            }
+            $restrictionsValue.passwordCredentials += $newItem
         }
 
-        $PinnedAppBarAppsValue = $instance.PinnedAppBarApps.Id
-        if ($instance.PinnedAppBarApps.Count -eq 0)
+        foreach ($keyCred in $instance.Restrictions.KeyCredentials)
         {
-            $PinnedAppBarAppsValue = @()
+            $newItem = @{
+                restrictForAppsCreatedAfterDateTime = $keyCred.RestrictForAppsCreatedAfterDateTime.ToString()
+                restrictionType                     = $keyCred.RestrictionType
+                state                               = $keyCred.State
+            }
+            if ($null -ne $keyCred.MaxLifetime)
+            {
+                $iso8601Duration = "P{0}DT{1}H{2}M{3}S" -f $keyCred.MaxLifetime.Days, $keyCred.MaxLifetime.Hours, $keyCred.MaxLifetime.Minutes, $keyCred.MaxLifetime.Seconds
+                $newItem.Add('maxLifetime', $iso8601Duration)
+            }
+            $restrictionsValue.keyCredentials += $newItem
         }
 
-        $PinnedMessageBarAppsValue = $instance.PinnedMessageBarApps.Id
-        if ($instance.PinnedMessageBarApps.Count -eq 0)
-        {
-            $PinnedMessageBarAppsValue = @()
-        }
-
-        Write-Verbose -Message "Found an instance with Identity {$Identity}"
         $results = @{
-            Identity              = $instance.Identity.Replace('Tag:', '')
+            DisplayName           = $instance.DisplayName
+            Id                    = $instance.Id
             Description           = $instance.Description
-            AppPresetList         = [Array]$AppPresetListValue
-            AppPresetMeetingList  = [Array]$AppPresetMeetingListValue
-            PinnedAppBarApps      = [Array]$PinnedAppBarAppsValue
-            PinnedMessageBarApps  = [Array]$PinnedMessageBarAppsValue
-            AllowUserPinning      = $instance.AllowUserPinning
-            AllowSideLoading      = $instance.AllowSideLoading
+            IsEnabled             = $instance.IsEnabled
+            Restrictions          = $restrictionsValue
             Ensure                = 'Present'
             Credential            = $Credential
             ApplicationId         = $ApplicationId
@@ -137,6 +148,7 @@ function Get-TargetResource
     }
     catch
     {
+        Write-Verbose -Message $_
         New-M365DSCLogEntry -Message 'Error retrieving data:' `
             -Exception $_ `
             -Source $($MyInvocation.MyCommand.Source) `
@@ -154,35 +166,23 @@ function Set-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $DisplayName,
 
         [Parameter()]
+        [System.String]
+        $Id,
+
+        [Parameter(Mandatory = $true)]
         [System.String]
         $Description,
 
         [Parameter()]
-        [System.String[]]
-        $AppPresetList,
-
-        [Parameter()]
-        [System.String[]]
-        $AppPresetMeetingList,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedAppBarApps,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedMessageBarApps,
-
-        [Parameter()]
         [System.Boolean]
-        $AllowUserPinning,
+        $IsEnabled,
 
         [Parameter()]
-        [System.Boolean]
-        $AllowSideLoading,
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $Restrictions,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -213,9 +213,6 @@ function Set-TargetResource
         [System.String[]]
         $AccessTokens
     )
-
-    New-M365DSCConnection -Workload 'MicrosoftTeams' `
-        -InboundParameters $PSBoundParameters | Out-Null
 
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
@@ -231,85 +228,60 @@ function Set-TargetResource
 
     $currentInstance = Get-TargetResource @PSBoundParameters
 
-    $PSBoundParameters.Remove('Ensure') | Out-Null
-    $PSBoundParameters.Remove('Credential') | Out-Null
-    $PSBoundParameters.Remove('ApplicationId') | Out-Null
-    $PSBoundParameters.Remove('ApplicationSecret') | Out-Null
-    $PSBoundParameters.Remove('TenantId') | Out-Null
-    $PSBoundParameters.Remove('CertificateThumbprint') | Out-Null
-    $PSBoundParameters.Remove('ManagedIdentity') | Out-Null
-    $PSBoundParameters.Remove('AccessTokens') | Out-Null
+    $setParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
 
-    $appPresetValues = @()
-    if ($null -ne $AppPresetList -and ([Array]$AppPresetList).Count -gt 0)
-    {
-        foreach ($appInstance in $AppPresetList)
-        {
-            $appPresetValues += [Microsoft.Teams.Policy.Administration.Cmdlets.Core.AppPreset]::New($appInstance)
-        }
+    $restrictionsValue = @{
+        passwordCredentials = @()
+        keyCredentials      = @()
     }
 
-    $appPresetMeetingValues = @()
-    if ($null -ne $AppPresetMeetingList -and ([Array]$AppPresetMeetingList).Count -gt 0)
+    foreach ($passwordCred in $Restrictions.PasswordCredentials)
     {
-        foreach ($appInstance in $AppPresetMeetingList)
-        {
-            $appPresetMeetingValues += [Microsoft.Teams.Policy.Administration.Cmdlets.Core.AppPresetMeeting]::New($appInstance)
+        $newItem = @{
+            restrictForAppsCreatedAfterDateTime = [System.DateTime]::Parse($passwordCred.RestrictForAppsCreatedAfterDateTime)
+            restrictionType                     = $passwordCred.RestrictionType
+            state                               = $passwordCred.State
         }
+        if ($null -ne $passwordCred.MaxLifetime)
+        {
+            $newItem.Add('maxLifetime', $passwordCred.MaxLifetime.ToString())
+        }
+        $restrictionsValue.passwordCredentials += $newItem
     }
 
-    $pinnedAppBarAppsValue = @()
-    if ($null -ne $PinnedAppBarApps -and ([Array]$PinnedAppBarApps).Count -gt 0)
+    foreach ($keyCred in $Restrictions.KeyCredentials)
     {
-        $i = 1
-        foreach ($appInstance in $PinnedAppBarApps)
-        {
-            $pinnedAppBarAppsValue += [Microsoft.Teams.Policy.Administration.Cmdlets.Core.PinnedApp]::New($appInstance, $i)
-            $i++
+        $newItem = @{
+            restrictForAppsCreatedAfterDateTime = [System.DateTime]::Parse($keyCred.RestrictForAppsCreatedAfterDateTime)
+            restrictionType                     = $keyCred.RestrictionType
+            state                               = $keyCred.State
         }
+        if ($null -ne $keyCred.MaxLifetime)
+        {
+            $newItem.Add('maxLifetime', $keyCred.MaxLifetime.ToString())
+        }
+        $restrictionsValue.keyCredentials += $newItem
     }
 
-    $pinnedMessageBarAppsValue = @()
-    if ($null -ne $PinnedMessageBarApps -and ([Array]$PinnedMessageBarApps).Count -gt 0)
-    {
-        $i = 1
-        foreach ($appInstance in $PinnedMessageBarApps)
-        {
-            $pinnedMessageBarAppsValue += [Microsoft.Teams.Policy.Administration.Cmdlets.Core.PinnedMessageBarApp]::New($appInstance, $i)
-            $i++
-        }
-    }
+    $setParameters.Restrictions = $restrictionsValue
 
+    # CREATE
     if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
-        $CreateParameters = ([Hashtable]$PSBoundParameters).Clone()
-        $CreateParameters.Remove('Verbose') | Out-Null
-        Write-Verbose -Message "Creating {$Identity} with Parameters:`r`n$(Convert-M365DscHashtableToString -Hashtable $CreateParameters)"
-
-        $CreateParameters.AppPresetList = $appPresetValues
-        $CreateParameters.AppPresetMeetingList = $appPresetMeetingValues
-        $CreateParameters.PinnedAppBarApps = $pinnedAppBarAppsValue
-        $CreateParameters.PinnedMessageBarApps = $pinnedMessageBarAppsValue
-
-        New-CsTeamsAppSetupPolicy @CreateParameters | Out-Null
+        Write-Verbose -Message "Creating new App Management Policy {$DisplayName} with:`r`n$(ConvertTo-Json $setParameters -Depth 10)"
+        New-MgBetaPolicyAppManagementPolicy @SetParameters
     }
+    # UPDATE
     elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
     {
-        $UpdateParameters = ([Hashtable]$PSBoundParameters).Clone()
-        $UpdateParameters.Remove('Verbose') | Out-Null
-        Write-Verbose -Message "Updating {$Identity}"
-
-        $UpdateParameters.AppPresetList = $appPresetValues
-        $UpdateParameters.AppPresetMeetingList = $appPresetMeetingValues
-        $UpdateParameters.PinnedAppBarApps = $pinnedAppBarAppsValue
-        $UpdateParameters.PinnedMessageBarApps = $pinnedMessageBarAppsValue
-
-        Set-CsTeamsAppSetupPolicy @UpdateParameters | Out-Null
+        Write-Verbose -Message "Updating App Management Policy {$DisplayName} with:`r`n$(ConvertTo-Json $setParameters -Depth 10)"
+        Update-MgBetaPolicyAppManagementPolicy @SetParameters -AppManagementPolicyId $currentInstance.Id
     }
+    # REMOVE
     elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Removing {$Identity}"
-        Remove-CsTeamsAppSetupPolicy -Identity $currentInstance.Identity
+        Write-Verbose -Message "Removing App Management Policy {$DisplayName}"
+        Remove-MgBetaPolicyAppManagementPolicy -AppManagementPolicyId $currentInstance.Id
     }
 }
 
@@ -321,35 +293,23 @@ function Test-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String]
-        $Identity,
+        $DisplayName,
 
         [Parameter()]
+        [System.String]
+        $Id,
+
+        [Parameter(Mandatory = $true)]
         [System.String]
         $Description,
 
         [Parameter()]
-        [System.String[]]
-        $AppPresetList,
-
-        [Parameter()]
-        [System.String[]]
-        $AppPresetMeetingList,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedAppBarApps,
-
-        [Parameter()]
-        [System.String[]]
-        $PinnedMessageBarApps,
-
-        [Parameter()]
         [System.Boolean]
-        $AllowUserPinning,
+        $IsEnabled,
 
         [Parameter()]
-        [System.Boolean]
-        $AllowSideLoading,
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $Restrictions,
 
         [Parameter()]
         [ValidateSet('Present', 'Absent')]
@@ -393,13 +353,39 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    Write-Verbose -Message "Testing configuration of {$Identity}"
-
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
+    $testTargetResource = $true
 
+    #Compare Cim instances
+    foreach ($key in $PSBoundParameters.Keys)
+    {
+        $source = $PSBoundParameters.$key
+        $target = $CurrentValues.$key
+        if ($null -ne $source -and $source.GetType().Name -like '*CimInstance*')
+        {
+            if (-not ($source.GetType().Name -eq 'CimInstance[]' -and $source.Count -eq 0))
+            {
+                $testResult = Compare-M365DSCComplexObject `
+                    -Source ($source) `
+                    -Target ($target)
 
-
+                if (-not $testResult)
+                {
+                    Write-Verbose "TestResult returned False for $source"
+                    $testTargetResource = $false
+                }
+                else
+                {
+                    $ValuesToCheck.Remove($key) | Out-Null
+                }
+            }
+            else
+            {
+                $ValuesToCheck.Remove($key) | Out-Null
+            }
+        }
+    }
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $ValuesToCheck)"
 
@@ -409,8 +395,12 @@ function Test-TargetResource
         -ValuesToCheck $ValuesToCheck.Keys
 
     Write-Verbose -Message "Test-TargetResource returned $testResult"
+    if (-not $TestResult)
+    {
+        $testTargetResource = $false
+    }
 
-    return $testResult
+    return $testTargetResource
 }
 
 function Export-TargetResource
@@ -448,7 +438,7 @@ function Export-TargetResource
         $AccessTokens
     )
 
-    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
         -InboundParameters $PSBoundParameters
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -465,30 +455,32 @@ function Export-TargetResource
 
     try
     {
-        [array]$getValue = Get-CsTeamsAppSetupPolicy -ErrorAction Stop
+        $Script:ExportMode = $true
+        [array] $Script:exportedInstances = Get-MgBetaPolicyAppManagementPolicy -ErrorAction Stop
 
         $i = 1
         $dscContent = ''
-        if ($getValue.Length -eq 0)
+        if ($Script:exportedInstances.Length -eq 0)
         {
-            Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark -CommitWrite
+            Write-M365DSCHost -Message $Global:M365DSCEmojiGreenCheckMark
         }
         else
         {
             Write-M365DSCHost -Message "`r`n" -DeferWrite
         }
-        foreach ($config in $getValue)
+        foreach ($config in $Script:exportedInstances)
         {
             if ($null -ne $Global:M365DSCExportResourceInstancesCount)
             {
                 $Global:M365DSCExportResourceInstancesCount++
             }
 
-            $displayedKey = $config.Identity
-            Write-M365DSCHost -Message "    |---[$i/$($getValue.Count)] $displayedKey" -DeferWrite
+            $displayedKey = $config.DisplayName
+            Write-M365DSCHost -Message "    |---[$i/$($Script:exportedInstances.Count)] $displayedKey" -DeferWrite
             $params = @{
-                Identity              = $config.Identity
-                Ensure                = 'Present'
+                DisplayName           = $config.DisplayName
+                Id                    = $config.Id
+                Description           = $config.Description
                 Credential            = $Credential
                 ApplicationId         = $ApplicationId
                 TenantId              = $TenantId
@@ -498,12 +490,45 @@ function Export-TargetResource
             }
 
             $Results = Get-TargetResource @Params
+            if ($null -ne $Results.Restrictions)
+            {
+                $complexMapping = @(
+                    @{
+                        Name            = 'Restrictions'
+                        CimInstanceName = 'AADAppManagementPolicyRestrictions'
+                        IsRequired      = $False
+                    }
+                    @{
+                        Name            = 'PasswordCredentials'
+                        CimInstanceName = 'AADAppManagementPolicyRestrictionsCredential'
+                        IsRequired      = $False
+                    }
+                    @{
+                        Name            = 'KeyCredentials'
+                        CimInstanceName = 'AADAppManagementPolicyRestrictionsCredential'
+                        IsRequired      = $False
+                    }
+                )
+                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString `
+                    -ComplexObject $Results.Restrictions `
+                    -CIMInstanceName 'AADAppManagementPolicyRestrictions' `
+                    -ComplexTypeMapping $complexMapping
 
+                if (-not [String]::IsNullOrWhiteSpace($complexTypeStringResult))
+                {
+                    $Results.Restrictions = $complexTypeStringResult
+                }
+                else
+                {
+                    $Results.Remove('Restrictions') | Out-Null
+                }
+            }
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `
                 -Results $Results `
-                -Credential $Credential
+                -Credential $Credential `
+                -NoEscape @('Restrictions', 'KeyCredentials', 'PasswordCredentials')
             $dscContent += $currentDSCBlock
             Save-M365DSCPartialExport -Content $currentDSCBlock `
                 -FileName $Global:PartialExportFileName